Moved layout components to layout module
