+++ src/gui.rs
@@ -2,27 +2,21 @@ use crate::collections::ordered_dag::prelude::*;
 use crate::collections::ordered_dag::{ChildrenWalk, PostOrderWalk};
 use specs::Entity;
 
-mod aabb;
 mod builder;
 mod draw;
 mod drawable;
 mod layout;
 mod mesh;
-mod placement;
-mod pos;
 mod proj;
 mod systems;
 mod widget;
 pub mod widgets;
 
-pub use aabb::*;
 pub use builder::*;
 pub use draw::*;
 pub use drawable::*;
 pub use layout::*;
 pub use mesh::*;
-pub use placement::*;
-pub use pos::*;
 pub use proj::*;
 pub use systems::*;
 pub use widget::*;
